chore: Style up page
